Refine position and layout of header and hero
Padding is being used in a such a way that the header and hero content
are equally pushed away from the left and right of the screen.
Hero text and the image are set to flex: 1 so they share the same
amount of space, although the image probably shouldn't be resizing.